jpql 실습 2 - 서브쿼리, 동적쿼리

+++ src/main/java/model/Department.java
@@ -12,11 +12,11 @@ public class Department {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name ="DEPARTMENT_ID" )
-    private  long deptid;
+    private  Long deptid;
     @Column(name="DEPARTMENT_NAME")
     private String dname;
     @Column(name="MANAGER_ID")
     private Integer mgrid;
     @Column(name="LOCATION_ID")
-    private Integer locid;
+    private Long locid;
 }
